fix: creating portable USBs
Removed all glob imports (from _ import *) to prevent uneeded loading and uneeded complexity (one of these imports was failing before)
Fixed a handful of syntax warnings
Added a couple else: raise ValueError to appease the linter and improve our error handling if that code path is triggered.
Added --contents-directory "." to the pyinstaller command so it doesn't create an _internal folder (new to the later versions of pyinstaller). This allows GUI_Base.ui to be found.